github_scrapers/requirements.txt
requests==2.32.3
pandas==2.2.3
SQLAlchemy==2.0.36
python-dotenv==1.0.1
discord.py==2.4.0
